pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "spaceinvaders"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = ["pygame-ce>=2.5.0"]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
